requirements.txt
requests
beautifulsoup4
pymongo
schedule
pytz
spacy
faiss-cpu
numpy
sentence-transformers
gunicorn
